requirements.txt
Flask~=1.1.2
Flask-RESTful==0.3.9
pymongo==4.1.1
dnspython==2.2.1
waitress~=2.1.1
python-decouple==3.6
mysql-connector-python~=8.0.29
markupsafe==2.0.1